requirements.txt
telebot
python-dotenv
